+++ apidaora/bodies.py
@@ -1,8 +1,7 @@
 import io
-from typing import IO, Optional, Type
+from typing import IO, Optional
 
 from dictdaora import DictDaora
-from jsondaora import jsondaora
 
 
 try:
@@ -11,15 +10,15 @@ except Exception:
     gzip = None  # type: ignore
 
 
-class _GZipFactory(DictDaora):  # type: ignore
-    mode: str
-    compresslevel: int
-    encoding: Optional[str]
-    errors: Optional[str]
-    newline: Optional[str]
+class GZipFactory(DictDaora):
+    mode: str = 'rb'
+    compresslevel: int = 9
+    encoding: Optional[str] = None
+    errors: Optional[str] = None
+    newline: Optional[str] = None
     value: bytes
 
-    def open(self) -> IO[bytes]:
+    def open(self) -> IO[str]:
         if self.value:
             return gzip.open(
                 io.BytesIO(self.value),
@@ -31,21 +30,3 @@ class _GZipFactory(DictDaora):  # type: ignore
             )
 
         raise ValueError(self.value)
-
-
-def gzip_body(
-    mode: str = 'rb',
-    compresslevel: int = 9,
-    encoding: Optional[str] = None,
-    errors: Optional[str] = None,
-    newline: Optional[str] = None,
-) -> Type[_GZipFactory]:
-    cls_attributes = dict(
-        mode=mode,
-        compresslevel=compresslevel,
-        encoding=encoding,
-        errors=errors,
-        newline=newline,
-    )
-    factory = type('GZipFactory', (_GZipFactory,), cls_attributes)
-    return jsondaora(factory)  # type: ignore