+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Add package typing
+
 ### 0.11.0 - 2019-10-22
 
  - Fix background task status code